+++ src/main/kotlin/com/github/iammohdzaki/jsonbox/actions/OpenJsonBoxWindowAction.kt
@@ -1,8 +1,8 @@
 package com.github.iammohdzaki.jsonbox.actions
 
 import com.github.iammohdzaki.jsonbox.dialog.JsonBoxDialog
+import com.github.iammohdzaki.jsonbox.icons.JsonBoxIcons
 import com.github.iammohdzaki.jsonbox.utils.JsonBoxBundle
-import com.intellij.icons.AllIcons
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
@@ -16,7 +16,7 @@ class OpenJsonBoxWindowAction : AnAction() {
 
     override fun update(e: AnActionEvent) {
         e.presentation.text = JsonBoxBundle.message("jsonbox.title")
-        e.presentation.icon = AllIcons.FileTypes.Json
+        e.presentation.icon = JsonBoxIcons.ToolbarIcon
         e.presentation.isEnabled = e.project != null
     }